refactor: split semantic_registry.py into domain-split package
Closes #702

The 783-line vultron/semantic_registry.py is replaced by a package
under vultron/semantic_registry/ split into domain sub-modules:

  report.py          — 6 report entries
  case.py            — 5 case-lifecycle entries
  actor.py           — 13 actor/role entries
  embargo.py         — 7 embargo entries
  sync.py            — 3 entries (CLOSE_CASE + 2 case-log sync)
  case_participant.py — 3 case-participant entries
  note.py            — 3 note entries
  status.py          — 4 status entries
  unknown.py         — 2 fallback sentinels

The assembly order in __init__.py preserves the exact original
dispatch sequence (46 entries, UNKNOWN last). All public API symbols
are re-exported from __init__ for zero-change caller compatibility.

Also updated:
- test/architecture/test_activity_factory_imports.py: ALLOWED_PREFIXES
  updated from the old .py path to the new package directory
- notes/semantic-registry.md: frontmatter and body references updated
- notes/README.md: description updated
- docs/reference/two-actor-demo-protocol.md: table entry updated
- docs/reference/codebase/CONCERNS.md: concern row updated to reflect
  completion
