+++ docs/reference/codebase/CONCERNS.md
@@ -9,7 +9,7 @@
 | high | Demo helpers still depend on `requests`, but `requests` is not declared in `project.dependencies` | `vultron/demo/utils.py`, `vultron/demo/helpers/verification.py`, `pyproject.toml` | Fresh non-dev installs may fail when demo or verification helpers make HTTP calls; `httpx` (`>=0.28.1`) is already a declared runtime dep | Migrate demo HTTP helpers from `requests` to `httpx` — tracked in GitHub issue #517 |
 | high | Legacy globals and cached DataLayers still influence the deployed FastAPI path despite newer `create_app()` isolation rules | `vultron/adapters/driving/fastapi/main.py`, `vultron/adapters/driving/fastapi/app.py`, `vultron/adapters/driving/fastapi/inbox_handler.py`, `vultron/adapters/driving/fastapi/outbox_handler.py`, `vultron/adapters/driven/datalayer_sqlite.py` | Cross-app state leakage or startup-order bugs can still appear in co-located or shared-process setups | Continue migrating the mounted app path toward per-app state and keep regression tests around actor isolation |
 | medium | Outbox draining is implemented as a 1-second polling loop over all actor DataLayers | `vultron/adapters/driving/fastapi/outbox_monitor.py` | Polling cost grows with actor count and can hide queue-depth issues | Consider event-driven wakeups or queue metrics if actor count grows |
-| medium | Several source files still exceed 500 lines and mix multiple responsibilities | `vultron/core/behaviors/case/nodes.py` (1502 lines), `vultron/adapters/driven/datalayer_sqlite.py` (1108 lines), `vultron/demo/scenario/two_actor_demo.py` (863 lines), `vultron/wire/as2/extractor.py` (821 lines), `vultron/core/use_cases/triggers/embargo.py` (902 lines), `vultron/semantic_registry.py` (783 lines) | Large, multi-responsibility files are harder to test, review, and maintain; several are also high-churn | Incrementally extract cohesive sub-modules; prioritize `nodes.py`, `extractor.py`, and `datalayer_sqlite.py` |
+| medium | Several source files still exceed 500 lines and mix multiple responsibilities | `vultron/core/behaviors/case/nodes.py` (1502 lines), `vultron/adapters/driven/datalayer_sqlite.py` (1108 lines), `vultron/demo/scenario/two_actor_demo.py` (863 lines), `vultron/wire/as2/extractor.py` (821 lines), `vultron/core/use_cases/triggers/embargo.py` (902 lines) | Large, multi-responsibility files are harder to test, review, and maintain; several are also high-churn | Incrementally extract cohesive sub-modules; prioritize `nodes.py`, `extractor.py`, and `datalayer_sqlite.py` |
 | medium | Current remote delivery uses plain HTTP POST via `DeliveryQueueAdapter`; signing and shared-inbox support remain stubbed | `vultron/adapters/driven/delivery_queue.py`, `vultron/adapters/driven/http_delivery.py`, `vultron/adapters/driving/shared_inbox.py` | Multi-party interoperability and security posture are limited until signed delivery paths exist | Treat signed delivery/shared inbox as explicit future work and avoid implying these flows are production-ready |
 
 ### 2) Technical Debt
@@ -20,7 +20,7 @@
 | `http_delivery.py` and `shared_inbox.py` are stubs with no implementation | Placeholder-driven design for future signed HTTP delivery and shared-inbox fan-out | `vultron/adapters/driven/http_delivery.py`, `vultron/adapters/driving/shared_inbox.py` | No production delivery path for signed or shared-inbox flows; callers may assume these are functional | Implement or clearly gate behind feature flags before any production use |
 | Documentation/code drift around architecture targets | Notes describe both target and current structure; new ASGIEmitter rules now live in a separate note | `notes/architecture-hexagonal.md`, `notes/architecture-ports.md`, `notes/architecture-adapters.md`, `notes/asgi-emitter.md`, `AGENTS.md` | New contributors may confuse target layout with what exists today | Keep onboarding docs explicitly split into "current reality" vs "target direction" and cite the ASGIEmitter note where relevant |
 | High churn in planning and guidance files | Planning/spec docs evolve rapidly during active development | `plan/BUILD_LEARNINGS.md`, `AGENTS.md` (79) in the latest scan window | Agent guidance and task context can go stale quickly | Treat `plan/` and guidance docs as volatile areas during onboarding; re-read before each session |
-| `vultron/semantic_registry.py` is a 783-line centralized dispatch table | Single module consolidates all message-type to handler wiring | `vultron/semantic_registry.py` | Any new message type requires editing this file; tight coupling can cause merge conflicts | Tracked in GitHub concern #508 — split into domain sub-modules under `vultron/semantic_registry/` |
+| `vultron/semantic_registry/` was a 783-line centralized dispatch table | Single module consolidated all message-type to handler wiring | `vultron/semantic_registry/` (refactored in #702) | Previously any new message type required editing one large file; now split into domain sub-modules | Split completed — see `vultron/semantic_registry/` package |
 
 ### 3) Security Concerns
