+++ vultron/semantic_registry/embargo.py
@@ -1,0 +1,104 @@
+"""Embargo-domain semantic registry entries."""
+
+#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
+#  - see Contributors.md for a full list of Contributors
+#  - see ContributionInstructions.md for information on how you can Contribute to this project
+#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
+#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
+#  with this Software or contact permission@sei.cmu.edu for full terms.
+#  Created, in part, with funding and support from the United States Government
+#  (see Acknowledgments file). This program may include and/or can make use of
+#  certain third party source code, object code, documentation and other files
+#  ("Third Party Software"). See LICENSE.md for more details.
+#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
+#  U.S. Patent and Trademark Office by Carnegie Mellon University
+
+from vultron.core.models.events.base import MessageSemantics
+from vultron.core.models.events.embargo import (
+    AcceptInviteToEmbargoOnCaseReceivedEvent,
+    AddEmbargoEventToCaseReceivedEvent,
+    AnnounceEmbargoEventToCaseReceivedEvent,
+    CreateEmbargoEventReceivedEvent,
+    InviteToEmbargoOnCaseReceivedEvent,
+    RejectInviteToEmbargoOnCaseReceivedEvent,
+    RemoveEmbargoEventFromCaseReceivedEvent,
+)
+from vultron.core.use_cases.received.embargo import (
+    AcceptInviteToEmbargoOnCaseReceivedUseCase,
+    AddEmbargoEventToCaseReceivedUseCase,
+    AnnounceEmbargoEventToCaseReceivedUseCase,
+    CreateEmbargoEventReceivedUseCase,
+    InviteToEmbargoOnCaseReceivedUseCase,
+    RejectInviteToEmbargoOnCaseReceivedUseCase,
+    RemoveEmbargoEventFromCaseReceivedUseCase,
+)
+from vultron.semantic_registry._entry import SemanticEntry
+from vultron.wire.as2.extractor import (
+    AcceptInviteToEmbargoOnCasePattern,
+    AddEmbargoEventToCasePattern,
+    AnnounceEmbargoEventToCasePattern,
+    CreateEmbargoEventPattern,
+    InviteToEmbargoOnCasePattern,
+    RejectInviteToEmbargoOnCasePattern,
+    RemoveEmbargoEventFromCasePattern,
+)
+from vultron.wire.as2.vocab.activities.embargo import (
+    _AddEmbargoToCaseActivity,
+    _AnnounceEmbargoActivity,
+    _EmAcceptEmbargoActivity,
+    _EmProposeEmbargoActivity,
+    _EmRejectEmbargoActivity,
+    _RemoveEmbargoFromCaseActivity,
+)
+
+ENTRIES: list[SemanticEntry] = [
+    SemanticEntry(
+        semantics=MessageSemantics.CREATE_EMBARGO_EVENT,
+        pattern=CreateEmbargoEventPattern,
+        event_class=CreateEmbargoEventReceivedEvent,
+        use_case_class=CreateEmbargoEventReceivedUseCase,
+    ),
+    SemanticEntry(
+        semantics=MessageSemantics.ADD_EMBARGO_EVENT_TO_CASE,
+        pattern=AddEmbargoEventToCasePattern,
+        event_class=AddEmbargoEventToCaseReceivedEvent,
+        use_case_class=AddEmbargoEventToCaseReceivedUseCase,
+        wire_activity_class=_AddEmbargoToCaseActivity,
+    ),
+    SemanticEntry(
+        semantics=MessageSemantics.REMOVE_EMBARGO_EVENT_FROM_CASE,
+        pattern=RemoveEmbargoEventFromCasePattern,
+        event_class=RemoveEmbargoEventFromCaseReceivedEvent,
+        use_case_class=RemoveEmbargoEventFromCaseReceivedUseCase,
+        wire_activity_class=_RemoveEmbargoFromCaseActivity,
+    ),
+    SemanticEntry(
+        semantics=MessageSemantics.ANNOUNCE_EMBARGO_EVENT_TO_CASE,
+        pattern=AnnounceEmbargoEventToCasePattern,
+        event_class=AnnounceEmbargoEventToCaseReceivedEvent,
+        use_case_class=AnnounceEmbargoEventToCaseReceivedUseCase,
+        wire_activity_class=_AnnounceEmbargoActivity,
+    ),
+    SemanticEntry(
+        semantics=MessageSemantics.INVITE_TO_EMBARGO_ON_CASE,
+        pattern=InviteToEmbargoOnCasePattern,
+        event_class=InviteToEmbargoOnCaseReceivedEvent,
+        use_case_class=InviteToEmbargoOnCaseReceivedUseCase,
+        wire_activity_class=_EmProposeEmbargoActivity,
+        include_activity=True,
+    ),
+    SemanticEntry(
+        semantics=MessageSemantics.ACCEPT_INVITE_TO_EMBARGO_ON_CASE,
+        pattern=AcceptInviteToEmbargoOnCasePattern,
+        event_class=AcceptInviteToEmbargoOnCaseReceivedEvent,
+        use_case_class=AcceptInviteToEmbargoOnCaseReceivedUseCase,
+        wire_activity_class=_EmAcceptEmbargoActivity,
+    ),
+    SemanticEntry(
+        semantics=MessageSemantics.REJECT_INVITE_TO_EMBARGO_ON_CASE,
+        pattern=RejectInviteToEmbargoOnCasePattern,
+        event_class=RejectInviteToEmbargoOnCaseReceivedEvent,
+        use_case_class=RejectInviteToEmbargoOnCaseReceivedUseCase,
+        wire_activity_class=_EmRejectEmbargoActivity,
+    ),
+]
